chore: rename @antv/webgpu-graph
