+++ packages/graph-gpu/src/traversal/bfs.ts
