+++ packages/graph-gpu/src/traversal/index.ts
