+++ packages/graph-gpu/src/types.ts
