+++ packages/graph-gpu/src/util.ts
