+++ site/benchmark/index.html
@@ -175,6 +175,11 @@
         <div class="console"></div>
         <span class="time"></span>
       </section>
+      <section id="@antv/graph-wasm">
+        <div class="title">@antv/graph-wasm</div>
+        <div class="console"></div>
+        <span class="time"></span>
+      </section>
     </div>
 
     <script type="module" src="./main.ts"></script>
