+++ site/benchmark/sssp.ts
@@ -1,5 +1,5 @@
 import { dijkstra } from "graphology-shortest-path";
-import { WebGPUGraph } from "../../packages/webgpu-graph/src";
+import { WebGPUGraph } from "../../packages/graph-gpu/src";
 
 export async function graphology(graph: any, source: string) {
   return dijkstra.singleSource(graph, source, "weight");
