+++ site/types.ts
@@ -2,4 +2,5 @@ export enum TestName {
   GRAPHOLOGY = "graphology",
   ANTV_ALGORITHM = "@antv/algorithm",
   ANTV_WEBGPU_GRAPH = "@antv/webgpu-graph",
+  ANTV_GRAPH_WASM = "@antv/graph-wasm",
 }
